Fixed some warnings
Fixed warnings included in Public GitHub build #15
- https://github.com/TcNobo/TcNo-Acc-Switcher/actions/runs/875268244

+++ TcNo-Acc-Switcher-Server/Pages/General/GeneralFuncs.cs
@@ -63,7 +63,13 @@ namespace TcNo_Acc_Switcher_Server.Pages.General
             bool canKill;
             if (AppSettings.Instance.CurrentlyElevated)
                 canKill = true; // Elevated process can kill most other processes.
-            else canKill = !ProcessHelper.IsProcessAdmin(processName); // If other process is admin, this process can't kill it (because it's not admin)
+            else
+            {
+                if (OperatingSystem.IsWindows())
+                    canKill = !ProcessHelper.IsProcessAdmin(processName); // If other process is admin, this process can't kill it (because it's not admin)
+                else 
+                    canKill = false;
+            }
 
             // Restart self as admin.
             if (!canKill) AppData.ActiveIJsRuntime.InvokeAsync<string>("ShowModal", "notice:RestartAsAdmin");
